add limit

+++ sheets/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render
 from django.conf import settings
 from django.http import HttpResponse
+from .utils.shortcuts import file_is_exist
 from bs4 import BeautifulSoup as bs
 import pandas as pd
 import os
@@ -13,41 +14,47 @@ def create_excel(request):
     """
     Scrap the Data From Wikipedia and place it in Excel
     """
-
     file_path = os.path.join(settings.BASE_DIR, "excel_files/data.xlsx")
 
-    domain = 'https://ar.wikipedia.org'
+    if file_is_exist(file_path):
 
-    url = "https://ar.wikipedia.org/wiki/%D9%82%D8%A7%D8%A6%D9%85%D8%A9_%D8%A3%D9%81%D8%B6%D9%84_%D9%85%D8%A6%D8%A9_%D8%B1%D9%88%D8%A7%D9%8A%D8%A9_%D8%B9%D8%B1%D8%A8%D9%8A%D8%A9"
+        return HttpResponse("File Is Already Exist")
 
-    response = requests.get(url)
+    else:
+        file_path = os.path.join(settings.BASE_DIR, "excel_files/data.xlsx")
 
-    soup = bs(response.content, features='html.parser')
+        domain = 'https://ar.wikipedia.org'
 
-    table = soup.select('table.wikitable')[0]
+        url = "https://ar.wikipedia.org/wiki/%D9%82%D8%A7%D8%A6%D9%85%D8%A9_%D8%A3%D9%81%D8%B6%D9%84_%D9%85%D8%A6%D8%A9_%D8%B1%D9%88%D8%A7%D9%8A%D8%A9_%D8%B9%D8%B1%D8%A8%D9%8A%D8%A9"
 
-    columns = [i.get_text(strip=True) for i in table.find_all("th")]
+        response = requests.get(url)
 
-    columns += ["رابط الكتاب", "رابط المؤلف", "رابط البلد"]
+        soup = bs(response.content, features='html.parser')
 
-    data = []
+        table = soup.select('table.wikitable')[0]
 
-    for tr in table.find("tbody").find_all("tr"):
-        cells = []
-        tds = tr.find_all('td')
-        link = []
+        columns = [i.get_text(strip=True) for i in table.find_all("th")]
 
-        for td in tds:
-            cells.append(td.get_text(strip=True))
-            if td.find('a'):
-                link.append(domain + td.find('a')['href'])
-        data.append(cells + link)
+        columns += ["رابط الكتاب", "رابط المؤلف", "رابط البلد"]
 
-    df = pd.DataFrame(data, columns=columns)
+        data = []
 
-    df.to_excel(file_path, index=False)
+        for tr in table.find("tbody").find_all("tr"):
+            cells = []
+            tds = tr.find_all('td')
+            link = []
 
-    return HttpResponse("Data Has Been Written Successfully")
+            for td in tds:
+                cells.append(td.get_text(strip=True))
+                if td.find('a'):
+                    link.append(domain + td.find('a')['href'])
+            data.append(cells + link)
+
+        df = pd.DataFrame(data, columns=columns)
+
+        df.to_excel(file_path, index=False)
+
+        return HttpResponse("Data Has Been Written Successfully")
 
 
 def create_google_sheet(request):
